+++ src/components/About/About.jsx
@@ -1,0 +1,10 @@
+
+const About = () => {
+    return (
+        <div>
+            About
+        </div>
+    )
+}
+
+export default About;
