--- src/components/Main.jsx
@@ -1,10 +1,0 @@
-
-
-const Main = (params) => {
-    return (
-        <main className="main">
-            Pictures
-        </main>
-    )
-}
-export default Main;
