+++ src/components/Main/Main.jsx
@@ -1,0 +1,10 @@
+import s from "./Main.module.css"
+
+const Main = (props) => {
+    return (
+        <main className={s.main}>
+            
+        </main>
+    )
+}
+export default Main;
