+++ src/index.css
@@ -24,49 +24,7 @@ body, input, select, textarea {
   height: 100%;
 } 
 
-nav {
-  grid-area: nav;
-  background-color: rgb(237, 247, 202);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  
-}
-nav a {
-  text-decoration: none;
-  padding: 0.5rem ;
-}
 
-nav img {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 5rem;
-  padding-right: 1rem;
-}
-
-.nav-container{
-  display: flex;
-  justify-content: start;
-  align-items: center;
-  padding-right: 30rem;
-}
-
-.nav-menu {
-  display: flex;
-  justify-content: end;
-  align-items: center;
-  padding: 0.5rem;
-}
-
-main {
+.main {
   grid-area: main;
-}
-
-footer {
-  grid-area: footer;
-  background-color: rgb(237, 247, 202);
-  display: flex;
-  justify-content: center;
-  align-items: center;
 }
